run from uber jar in cwd

+++ grind1.sh
@@ -22,5 +22,16 @@ echo "YAML2: ${IN2}" 1>&2;
 #./grind1.sh yaml yaml2 -sense "breathe%2:29:00::"
 #./grind1.sh yaml yaml2 1740-v
 
-jar=target/yaml2wndb-2.4.0-uber.jar
+jar=yaml2wndb-2.4.0-uber.jar
+if [ ! -e "${jar}" ]; then
+  if [ ! -e "target/${jar}" ]; then
+    echo "Non existing uber jar" >&2
+    exit 1
+    fi
+  ln -s "target/${jar}"
+  fi
+if [ ! -e "${jar}" ]; then
+  echo "Non existing uber jar" >&2
+  exit 2
+  fi
 java -ea -cp "${jar}" org.oewntk.grind.yaml2wndb.Grind1 "${IN}" "${IN2}" $3 $4 $5
